requirements.txt
django==2.2
mysqlclient==1.4.6
django-crispy-forms==1.7.2
django-mptt==0.10.0
ansible==2.4.2.0
django-simple-captcha==0.5.6
paramiko==2.4.1
celery==4.3.0
redis==3.2.1
django-celery-results==1.1.2
django_celery_beat==1.5.0
pymysql==0.9.3
requests==2.22.0
raven==6.10.0
gunicorn==19.9.0
djangorestframework==3.10.3
markdown==3.1.1
django-filter==2.2.0
psutil==5.6.7
mongoengine